+++ runtime/services/services.go
@@ -22,6 +22,7 @@ import (
 	"github.com/hypermodeinc/modus/runtime/logger"
 	"github.com/hypermodeinc/modus/runtime/manifestdata"
 	"github.com/hypermodeinc/modus/runtime/middleware"
+	"github.com/hypermodeinc/modus/runtime/modusdb"
 	"github.com/hypermodeinc/modus/runtime/pluginmanager"
 	"github.com/hypermodeinc/modus/runtime/secrets"
 	"github.com/hypermodeinc/modus/runtime/sqlclient"
@@ -52,6 +53,7 @@ func Start(ctx context.Context) context.Context {
 	secrets.Initialize(ctx)
 	storage.Initialize(ctx)
 	db.Initialize(ctx)
+	modusdb.Init(ctx)
 	collections.Initialize(ctx)
 	manifestdata.MonitorManifestFile(ctx)
 	envfiles.MonitorEnvFiles(ctx)
@@ -78,4 +80,5 @@ func Stop(ctx context.Context) {
 	dgraphclient.ShutdownConns()
 	logger.Close()
 	db.Stop(ctx)
+	modusdb.Close(ctx)
 }